+++ src/lib.rs
@@ -3,7 +3,7 @@ extern crate lalrpop_util;
 
 mod ast;
 pub mod validate;
-pub mod libpython;
+pub mod python_bindings;
 
 lalrpop_mod!(pub valve_grammar);
